modulo4 - TestandoAPI/requirements.txt
astroid==2.3.3
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.12
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jsonpath==0.82
lazy-object-proxy==1.4.3
Markdown==3.2.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.11.2
zipp==3.1.0
